docker/requirements.txt
PyYAML==3.12
zipfile2==0.0.12
